plat/arm: scp: Move to drivers/ folder
