+++ drivers/arm/css/scp/css_pm_scmi.c
@@ -9,13 +9,12 @@
 
 #include <arch_helpers.h>
 #include <common/debug.h>
+#include <drivers/arm/css/css_scp.h>
 #include <drivers/arm/css/scmi.h>
 #include <plat/arm/common/plat_arm.h>
 #include <plat/arm/css/common/css_pm.h>
 #include <plat/common/platform.h>
 #include <platform_def.h>
-
-#include "css_scp.h"
 
 /*
  * This file implements the SCP helper functions using SCMI protocol.
