+++ drivers/arm/css/scp/css_pm_scpi.c
@@ -8,11 +8,10 @@
 
 #include <arch_helpers.h>
 #include <common/debug.h>
+#include <drivers/arm/css/css_scp.h>
 #include <drivers/arm/css/css_scpi.h>
 #include <plat/arm/common/plat_arm.h>
 #include <plat/arm/css/common/css_pm.h>
-
-#include "css_scp.h"
 
 /*
  * This file implements the SCP power management functions using SCPI protocol.
